Do not error out if a server replies with M=0 in blockwise transfer
Servers are allowed to reply with M=0 to intermediate blocks for
requests transferred using block transfer. From the RFC:

> If the M bit was set in the request, the server can choose whether to
> act on each block separately, with no memory, or whether to handle the
> request for the entire body atomically, or any mix of the two.

The old code errored out if a server decided to act to each block
seperately. The new behaviour is to only error out if the response is
not a 2.xx code and continue otherwise. The deferred still only succeeds
after every packet has been successfully transmitted.